src/requirements.txt
argparse
pandas
configparser
librosa
numpy
soundfile
jinja2
boto3
scipy
asyncio
azure-core==1.13.0
azure-storage-blob==12.8.1
